Move shoot addon nginx-ingress tests to the common package
